flask/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
ujson
gunicorn==23.0.0
